GameHandler, gamecreation, console functionality and client interaction.
Type commant "create game" once you start the server to create game room and be able to accept clients joining, from the client side just  hit play and it will join you to the game and start playing.

Immediate next step, broadcast player creation in game to other players and start sending the location.

Next steps, join a game with the teammate that you are in lobby with and start creating the chain mechanics.